requirements.txt
# This file may be used to create an environment using:
# platform: linux-64
# created-by: conda 24.9.2
numpy=1.26.0=pypi_0
opencv-python-headless=4.10.0.84
albumentations=0.4.6
pandas=2.2.3
pcre2=10.42
pillow=10.4.0
python=3.12.2
scikit-image
scikit-learn=1.5.2
scipy=1.14.1
pytorch-cuda=12.4
pytorch=2.5.1
tensorboard=2.18.0
tqdm=4.66.5
yaml=0.2.5
